--- a/src/GraphQL/Execution/ParallelExecutionStrategy.cs
+++ b/src/GraphQL/Execution/ParallelExecutionStrategy.cs
@@ -14,7 +14,7 @@ namespace GraphQL.Execution
             var pendingNodes = new Queue<ExecutionNode>();
             pendingNodes.Enqueue(rootNode);
 
-            var currentTasks = new List<Task<ExecutionNode>>();
+            var currentTasks = new List<(Task, ExecutionNode)>();
             while (pendingNodes.Count > 0)
             {
                 // Start executing pending nodes, while limiting the maximum number of parallel executed nodes to the set limit
@@ -26,10 +26,12 @@ namespace GraphQL.Execution
                     var pendingNodeTask = ExecuteNodeAsync(context, pendingNode);
                     if (pendingNodeTask.IsCompleted)
                     {
+                        //throw any caught exceptions
+                        await pendingNodeTask;
+
                         // Node completed synchronously, so no need to add it to the list of currently executing nodes
                         // instead add any child nodes to the pendingNodes queue directly here
-                        var result = await pendingNodeTask;
-                        if (result is IParentExecutionNode parentExecutionNode)
+                        if (pendingNode is IParentExecutionNode parentExecutionNode)
                         {
                             foreach (var childNode in parentExecutionNode.GetChildNodes())
                             {
@@ -40,7 +42,7 @@ namespace GraphQL.Execution
                     else
                     {
                         // Node is actually asynchronous, so add it to the list of current tasks being executed in parallel
-                        currentTasks.Add(pendingNodeTask);
+                        currentTasks.Add((pendingNodeTask, pendingNode));
                     }
 
                 }
@@ -49,17 +51,18 @@ namespace GraphQL.Execution
                     .ConfigureAwait(false);
 
                 // Await tasks for this execution step
-                var completedNodes = await Task.WhenAll(currentTasks)
+                await Task.WhenAll(currentTasks.Select(x => x.Item1))
                     .ConfigureAwait(false);
-                currentTasks.Clear();
 
                 // Add child nodes to pending nodes to execute the next level in parallel
-                foreach (var node in completedNodes)
+                foreach (var node in currentTasks.Select(x => x.Item2))
                     if (node is IParentExecutionNode p)
                 {
                     foreach (var childNode in p.GetChildNodes())
                         pendingNodes.Enqueue(childNode);
                 }
+
+                currentTasks.Clear();
             }
         }
     }