This small change reduces an allocation per _synchronous_ resolver.  Note that some of the performance benefits may be lost with the updated DataLoader code, unfortunately.

```
// * Summary *

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18362
Intel Core i7-7740X CPU 4.30GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.1.101
  [Host]     : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X64 RyuJIT
  DefaultJob : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X64 RyuJIT

*** BEFORE ***

|        Method |        Mean |    Error |   StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------- |------------:|---------:|---------:|---------:|--------:|------:|----------:|
| Introspection | 1,509.92 us | 9.103 us | 8.070 us | 207.0313 | 76.1719 |     - | 1050.9 KB |
|          Hero |    31.69 us | 0.201 us | 0.178 us |   6.2256 |       - |     - |  25.55 KB |

*** AFTER ***

|        Method |        Mean |     Error |   StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------- |------------:|----------:|---------:|---------:|--------:|------:|----------:|
| Introspection | 1,478.62 us | 10.099 us | 9.446 us | 187.5000 | 76.1719 |     - | 943.65 KB |
|          Hero |    31.95 us |  0.320 us | 0.299 us |   6.1646 |       - |     - |   25.3 KB |
```